feat(treasury): Layer 3 — daily cron + admin Reconciliation panel (web) + mobile sentinel
- pgTAP fix: cast \gset delta to bigint so is() type-matches (was failing 6/6).
- config.toml: reconcile-ledger verify_jwt=false (cron via RECON_CRON_SECRET;
  admin Run-now via its own requireUser).
- .github/workflows/reconcile-ledger-cron.yml: daily 06:00 UTC curl of the edge
  function with x-cron-secret; job FAILS on shortfall (alerting). Skips if
  RECON_FUNCTION_URL secret absent.
- web data/treasury.ts: ReconciliationRunRow + fetchReconciliationRuns (admin RLS).
- web actions/treasury.ts: runReconciliation() invokes the edge fn, banners ok/shortfall.
- web treasury/page.tsx: Reconciliation panel (latest + history + Run now + drift),
  reusing the existing dark section/pill design system.
- mobile (admin)/financial.tsx: compact read-only reconciliation sentinel card
  (solvent/shortfall/snapshot) using C.* theme tokens; no banned separators.

Verified: web tsc clean on touched files; mobile scoped tsc clean; cron YAML valid;
schema-drift guard clean for new refs (pre-existing work_sessions flag unrelated).

+++ supabase/tests/reconciliation_test.sql
@@ -43,8 +43,8 @@ set local request.jwt.claims to '{"sub":"33333333-3333-3333-3333-333333333333","
 select (public.admin_ledger_snapshot()->>'liabilities_cents')::bigint as after \gset
 select (public.admin_ledger_snapshot()->>'wallets_cents')::bigint    as w_after \gset
 
-select is(:after - :before,   43500::bigint, 'snapshot liabilities reflect the +43500c we added');
-select is(:w_after - :w_before, 17500::bigint, 'wallet bucket reflects +17500c (dollars→cents)');
+select is((:after - :before)::bigint,    43500::bigint, 'snapshot liabilities reflect the +43500c we added');
+select is((:w_after - :w_before)::bigint, 17500::bigint, 'wallet bucket reflects +17500c (dollars→cents)');
 
 -- ── Drift logic: Stripe figure computed inline vs live liabilities ───────────
 select is(
